Add notebook for investigation of variant annotations format
Can't ensure that those are all the possible ones, but ABraOM seems to
have grouped two types into the same, since it brings 14 unique annota-
tions for the same variant searches performed in gnomad2 and gnomad3.

Gnomad 2 and 3 apparently use the same strings to refer to the types of
variants in the Annotation column.